+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 * Passing and invalid keyword argument to `build_command_string()` raises a `KeyError`.
 
+### 🔧 Fixed
+
+* Fix handling of boolean arguments.
+
 
 ## 0.1.0b3 - Mar 9, 2023